Move Ably usage into Internal module, and update access modifiers
