+++ Sources/AblyAssetTrackingCore/Model/TokenRequest.swift
@@ -1,23 +1,10 @@
 import Foundation
-import Ably
 
 public class TokenRequest: NSObject, Codable {
-    let keyName: String
-    let clientId: String
-    let capability: String
-    let timestamp: Int
-    let nonce: String
-    let mac: String
-}
-
-extension TokenRequest {
-    /**
-     Convert into ART (Ably Runtime representation)
-     */
-    func toARTTokenRequest() -> ARTTokenRequest {
-        let artTokenParams = ARTTokenParams(clientId: clientId, nonce: nonce)
-        artTokenParams.timestamp = Date(timeIntervalSince1970: Double(timestamp) / 1000)
-        artTokenParams.capability = capability
-        return ARTTokenRequest(tokenParams: artTokenParams, keyName: keyName, nonce: nonce, mac: mac)
-    }
+    public let keyName: String
+    public let clientId: String
+    public let capability: String
+    public let timestamp: Int
+    public let nonce: String
+    public let mac: String
 }
