+++ Sources/AblyAssetTrackingInternal/AblyExtensions/TokenDetails+Extensions.swift
@@ -1,8 +1,9 @@
-//
-//  File.swift
-//  
-//
-//  Created by Ben Butterworth on 19/06/2021.
-//
-
 import Foundation
+import Ably
+import AblyAssetTrackingCore
+
+extension TokenDetails {
+    public func toARTTokenDetails() -> ARTTokenDetails {
+        return ARTTokenDetails(token: token, expires: expires, issued: issued, capability: capability, clientId: clientId)
+    }
+}
